--- frontend/deploy.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-npm run generate
-aws s3 sync dist s3://pm3wmz3x4or6oolruyma
