+++ frontend/middleware/auth.global.js
