+++ frontend/plugins/vuetify.ts
