--- frontend/store/page.js
@@ -1,9 +1,0 @@
-export const state = () => ({
-  title: ''
-})
-
-export const mutations = {
-  setTitle(state, t) {
-    state.title = t  
-  }
-}
