--- frontend3/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "nuxt-app",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare"
-  },
-  "dependencies": {
-    "@mdi/font": "^7.4.47",
-    "nuxt": "^3.13.0",
-    "vue": "latest",
-    "vue-router": "latest"
-  },
-  "devDependencies": {
-    "vite-plugin-vuetify": "^2.0.4",
-    "vuetify": "^3.7.1"
-  }
-}
