--- frontend3/pages/logout.vue
@@ -1,7 +1,0 @@
-<script setup>
-  // composables
-  const auth = useAuth()
-  auth.value.authenticated = false
-  localStorage.clear()
-  await navigateTo('/login')
-</script>
